Refactor MiniDSPController component + Linting
- Updated .eslintrc.json to include additional rules for code quality, such as max-lines, max-lines-per-function, complexity, max-depth, and max-nested-callbacks.
- Refactored MiniDSPController component to improve structure and readability by consolidating input/output channel management into dedicated components (InputChannel, OutputChannel).
- Simplified the connection handling and UI layout for better user experience.
- Improved state management and API interaction for fetching and updating device configurations.

These changes enhance code maintainability and improve the overall functionality of the MiniDSPController.
